+++ packages/form/src/index.ts
@@ -1,15 +1,3 @@
 export * from './components/tj-form/tj-form';
-export * from './lib/Scope/FormDataContainer';
-export { createFormScope, createScope, defineArray, defineScope } from './lib/Scope/scope';
-export type {
-  ArrayDefinition,
-  Scope,
-  ScopeArray,
-  ScopeDefinition,
-  ScopeListener,
-  ScopeValue,
-  ScopeValueDefinition,
-  ScopeValueInput,
-} from './lib/Scope/scope-types';
-export { ScopeProxy } from './lib/Scope/ScopeProxy';
 export * from './lib/TjFormRegistry';
+export * from './plugins/EnterNextPlugin';