+++ packages/form/src/lib/Scope/ScopeArray.ts
@@ -1,1 +1,0 @@
-export { ScopeArrayRuntime as ScopeArray } from '@trunkjs/scope';